+++ Mail2Bug/WorkItemManagement/TFSWorkItemManager.cs
@@ -200,7 +200,7 @@ namespace Mail2Bug.WorkItemManagement
 
         /// <param name="values">The list of fields and their desired values to apply to the work item</param>
         /// <returns>Work item ID of the newly created work item</returns>
-        public int CreateWorkItem(Dictionary<string, string> values)
+        public long CreateWorkItem(Dictionary<string, string> values)
         {
             if (values == null)
             {
